+++ api/src/main/java/org/allaymc/api/command/CommandRegistry.java
@@ -1,9 +1,7 @@
 package org.allaymc.api.command;
 
-import org.allaymc.api.entity.interfaces.EntityPlayer;
 import org.allaymc.api.registry.RegistryLoader;
 import org.allaymc.api.registry.SimpleMappedRegistry;
-import org.cloudburstmc.protocol.bedrock.packet.AvailableCommandsPacket;
 
 import java.util.Map;
 
@@ -26,7 +24,6 @@ public abstract class CommandRegistry extends SimpleMappedRegistry<String, Comma
      * Method to unregister a command.
      *
      * @param name The name of the command to unregister
-     *
      * @return The unregistered command
      */
     public abstract Command unregister(String name);
@@ -36,17 +33,27 @@ public abstract class CommandRegistry extends SimpleMappedRegistry<String, Comma
      *
      * @param sender The command sender
      * @param cmd    The command, without the slash (/)
-     *
      * @return The command execution result
      */
     public abstract CommandResult execute(CommandSender sender, String cmd);
 
     /**
-     * Method to encode an AvailableCommandsPacket for a given player.
+     * Finds a command by its name or alias.
+     * <p>
+     * This method first searches for a command matching the given name or alias directly in the registry. If
+     * no match is found, it then searches through the aliases of all registered commands to find a match.
      *
-     * @param player The player to encode the packet for
-     *
-     * @return The encoded AvailableCommandsPacket
+     * @param nameOrAlias the name or alias of the command to find
+     * @return the command matching the given name or alias, or {@code null} if no command is found
      */
-    public abstract AvailableCommandsPacket encodeAvailableCommandsPacketFor(EntityPlayer player);
+    public Command findCommand(String nameOrAlias) {
+        var result = this.get(nameOrAlias);
+        if (result != null) {
+            return result;
+        }
+
+        return this.getContent().values().stream()
+                .filter(command -> command.getAliases().contains(nameOrAlias))
+                .findFirst().orElse(null);
+    }
 }