+++ api/src/main/java/org/allaymc/api/item/data/ArmorType.java
@@ -1,6 +1,8 @@
 package org.allaymc.api.item.data;
 
 /**
+ * Represents the type of armor item.
+ *
  * @author daoge_cmd
  */
 public enum ArmorType {