+++ server/src/main/java/org/allaymc/server/command/defaults/MeCommand.java
@@ -4,7 +4,6 @@ import org.allaymc.api.command.tree.CommandTree;
 import org.allaymc.api.message.TrKeys;
 import org.allaymc.api.permission.PermissionGroups;
 import org.allaymc.api.server.Server;
-import org.cloudburstmc.protocol.bedrock.data.command.CommandData;
 
 /**
  * @author daoge_cmd
@@ -12,7 +11,6 @@ import org.cloudburstmc.protocol.bedrock.data.command.CommandData;
 public class MeCommand extends VanillaCommand {
     public MeCommand() {
         super("me", TrKeys.MC_COMMANDS_ME_DESCRIPTION);
-        flags.add(CommandData.Flag.MESSAGE_TYPE);
         getPermissions().forEach(PermissionGroups.MEMBER::addPermission);
     }