+++ server/src/main/java/org/allaymc/server/command/defaults/VanillaCommand.java
@@ -1,17 +1,17 @@
 package org.allaymc.server.command.defaults;
 
-import org.allaymc.api.command.SimpleCommand;
+import org.allaymc.api.command.Command;
 import org.allaymc.api.permission.Permission;
-
-import java.util.List;
 
 /**
  * @author IWareQ
  */
-public abstract class VanillaCommand extends SimpleCommand {
+public abstract class VanillaCommand extends Command {
     public static final String ALLAY_BUILD_IN_COMMAND_PERMISSION_PREFIX = "allay.command.";
 
     public VanillaCommand(String name, String description) {
-        super(name, description, List.of(Permission.createForCommand(name, ALLAY_BUILD_IN_COMMAND_PERMISSION_PREFIX + name)));
+        super(name, description);
+        this.permissions.clear();
+        this.permissions.add(Permission.createForCommand(name, ALLAY_BUILD_IN_COMMAND_PERMISSION_PREFIX + name));
     }
 }