+++ server/src/main/java/org/allaymc/server/command/tree/node/CmdNode.java
@@ -3,7 +3,6 @@ package org.allaymc.server.command.tree.node;
 import org.allaymc.api.command.tree.CommandNode;
 import org.cloudburstmc.protocol.bedrock.data.command.CommandParam;
 import org.cloudburstmc.protocol.bedrock.data.command.CommandParamData;
-import org.cloudburstmc.protocol.bedrock.data.command.CommandParamOption;
 
 /**
  * @author daoge_cmd
@@ -11,7 +10,6 @@ import org.cloudburstmc.protocol.bedrock.data.command.CommandParamOption;
 public class CmdNode extends MessageNode {
     public CmdNode(String name, CommandNode parent, String defaultValue) {
         super(name, parent, defaultValue);
-        paramOptions.add(CommandParamOption.HAS_SEMANTIC_CONSTRAINT);
     }
 
     @Override