+++ server/src/main/java/org/allaymc/server/command/tree/node/KeyNode.java
@@ -1,7 +1,6 @@
 package org.allaymc.server.command.tree.node;
 
 import lombok.extern.slf4j.Slf4j;
-import org.allaymc.api.command.tree.BaseNode;
 import org.allaymc.api.command.tree.CommandContext;
 import org.allaymc.api.command.tree.CommandNode;
 import org.cloudburstmc.protocol.bedrock.data.command.CommandEnumData;